Change Table API back to columns array again (instead of Columns component children)
Main reason for that is to support automatic type inference for row argument in render method.

One neat thing that was possible using Column component was conditional hiding using {false && <Column syntax.
As replacement for that I've added a visible flag.

This is obviously absolutely incompatible and would require relatively big refactoring of all tables.