Add persistedState used by table states to keep filters during unmount
Problem appears when state (filter, paging, sort) is created inside a Stack page which is unmounted
when showing detail form. On remount (when user goes back) the filters are reset to default values.

This is now worked around by creating a persistent state id at stack level and passing it down to the
table. usePersistedStore backs up the value on unmount and reuses that value on remount.

Persisted might be missleading as it is not persisted across page reloads, only across mount/unmounts.

However we could probably want to persist (optionally) to localstorage to support also that (if needed)

Also add a story showing the problem, plus another story for an alternative solution (where all states are at stack level)