New MasterLayout with theme-support
The empty toolbar component is no longer needed for content-spacing,
this is now done with padding-top on the content-wrapper.

The grid-containers are not required because there are no grid-items.

The width-transition of the header was removed because the
header-width does not change.

The `headerHeight` prop is passed into the menu-context, so it only
has to be set on MasterLayout. It cannot be passed to the menu as a
prop because the menu-component passed from the admin-project should
not be required to accept that prop and pass it into the menu-component
from comet-admin.