Style components with comet-design by default
Styling the components with a generic design and then overriding the
styles with the comet-theme and maintaining both styles is probably
unnecessary. The generic style would probably be used rarely, if ever.
In most cases it will be just as easy to customize the comet-styles as
it would be if only the generic styled needed to be customized.