Adjustments and styling for tabs
Remove `AppBar` from around the `MuiTabs`, it is not required in the
comet-design (and any other of our current project-specific designs)
and would therefore have to be "styled away". It probably makes more
sense, to add a `background-color` and `box-shadow` to the `MuiTabs`
directly if any design requires it.

Remove unnecessary `tabLabel` prop, it was an alternative to label,
that did the exact same thing.