Replace FormPaper and some instances of Paper with Card
The FormPaper component was used inside forms to add the "paper-style" and some spacing for the form-fields inside.
For the same effect, outside of forms, a Paper component was used and the spacing was added manually for every instance.

This can be simplified, by adding a Card component instead of the Paper and CardContent inside, which adds the spacing for the inner content or form-fields.
For the "paper-style" without a spacing the normal Paper component should still be used.