Add prop to render FormSection without Typography
Previously, only when passing a string into the `title` prop,
the title would be wrapped by a Typography.

The thought behind this, was that a generic style would be added to the
title-text by default and when a custom component (e.g. Typography) was
needed that could simply be passed into the `title` prop and the
generic Typography would not be rendered.

The issue with this method was, that when passing in a component that
simply returned a text (e.g. FormattedMessage), the title would not
have the generic style because the type was not a string.