Prop for controlling margin of FieldContainer
This adds the ability to control if a FieldContainer should have a
margin at the end. In rare cases it may be necessary to remove margins
of a FieldContainer or add a margin to the last element of a form.

Usually there should be a margin from one FieldContainer to the next.
The last item usually should not have a margin, to prevent too large
spacing, e.g. when using Fields in a Card or FormSection.