Remove input-container min-width
The original purpose of the min-width was primarily to prevent selects
with no value from being very tiny.

Adding the min-width caused selects with short values to be
unnecessarily large and to break layouts. A better solution would be,
to make sure every select, that is not full-width has a default value
or placeholder.