Create new date-time package with DatePicker
Created a new date-time package to contain all date and time related
components, for use in forms, such as DatePicker, DateRangePicker,
TimePicker, DateTimePicker.

Implemented the DatePicker and FinalFormDatePicker, for now, using the
default styling, provided by 'react-date-range' and without allowing
text-input by the user.

Created the InputWithPopper component in the admin package to use as a
base for the DatePicker and other future pickers. It renders the input
and handles the opening and closing logic of the popper, preventing
inconsistencies in the behaviour of the different picker components.