Handle Prompt-Actions in DirtyHandler first
Now three options (Cancel, Discard, Save) exist in Prompt (Dialog
which will be shown when user wants to navigate away from a page
with a Dirty state). To handle the user action DirtyHandler registers
a callback into the RouterContext. The result of this callback
(allowTransition) is used for the react-router callback.