Implement CopyToClipboardButton & ContentWithHoverActions
`ContentWithHoverActions` must contain its content as `children` and
cannot simply be rendered as a sibling that overlays the content. The
reason is that overlaying the content would prevent the ability to
interact with the content, e.g., using the cursor to select part of
the content. Using `pointer-events: none` is not a viable solution, as
this would prevent the ability to detect the hover.