Fix merging of styleOverrides
A component's `styleOverrides` can be either an object or a function
that returns an object. Until recently, the `styleOverrides` inside
`@comet/admin-theme` and likely those defined in the application were
all objects. Therefore merging the two (the comet theme options and
the theme options defined in an application) always worked.

When updating to MUI 5, some `styleOverrides` inside
`@comet/admin-theme` were changed to functions. This would cause the
application-defined `styleOverrides` to replace the function instead
of merging them and removing the comet styles.

With `mergeOverrideStyles`, a new overrides function is created,
which calls the overrides-functions and merges their resulting objects.