Fix z-index issues of MasterLayout & MuiDrawer
The header inside MasterLayout needs to be above the menu to prevent
the content, e.g., MenuButton in the header, from being overlayed and
therefore not clickable. This was fixed in #694, which then caused the
MasterLayout-header to overlay all drawers (MuiDrawer).

Also, MuiDrawer-root should not always have a z-index, as it is not
always rendered as a modal, e.g., when using the variant "permanent".

By not using a custom theme value for the z-index of MuiDrawer-root and
instead only changing the z-index of components directly used inside
MasterLayout, these problems should be fixed permanently.