Workaround to force story re-render on state-change
By default, the story is not re-rendered when only the time of a state
is changed. This is likely a bug with storybook as this doesn't seem
to be reproducible without storybook.

By defining a new "React" variable with a custom set-function, the
existing stories work without any change, which is necessary so they
can be used as copy-paste examples.