Prevent flickering of StackBreadcrumbs
by only passing the breadcrumbs array's relevant values (as number &
string) into useEffect's dependencies.

Previously the array was passed in directly, which caused a
re-calculation of the item widths whenever the array was re-initialized,
even when the values did not change. This is because JS compares the
memory addresses instead of the values when comparing arrays & objects.