Fix dependencies so that eg mikro-orm/core is only installed once
Problem can be that mikro-orm/core has an optional peerDepdenency on sqlite. Now if in a monorepo one package
installs sqlite and one doesn't that resulsts in two different core installs (with different hashs).
This is problematic for nestjs DI and can be solved by installing sqlite also in demo.