Add spacing by default between Tabs and TabContent
Previously the spacing would have to be added manually for every tab,
which tended to be forgotten.

Now, with the default spacing, the tab content looks good
automatically without needing to add additional styling.

The breadcrumbs are an exception, as, according to the design, they
should be "stuck" to the tabs rather than be part of the tab content.