fix(createOffsetLimitPagingAction): calculate currentPage based on current offset and limit
This prevents the page and offset from getting out of sync. Getting out of sync can, for example, happen when using useTableQueryPaging to get a PagingApi and setting the inital value to 20 (anything else than 0 actually). In useTableQueryPaging, the currentPage is always set to 1. So, an initial value of 20 causes the frontend to show items 21 - 30 on page 1. Meaning, it's not possible for the user to view items 1 - 20.