Validate dam files created thru import mutation correctly
For that I had to refactor the DamUploadFileInterceptor into an own FileValidationService that gets also
called by the import mutation.

The FileValidationInterceptor that was implemented as own Interceptor I also refactored into the FileValidationService, as
using the Interceptor directly is not possible for the import mutation. Also I decided to put all validations into one Service
as else it might lead to one getting forgotten.