Ignore field resolvers in changes checker interceptor (#1336)
Scoped mutations would cause a change for all scopes when querying a
field resolver with `fieldResolverEnhancers: ["interceptors"]` enabled
in the `GraphQLModule` config. This is due to using `operator.operator`
to check whether an operation is a mutation. To fix this, we now use
`parentType.name`, as recommended in the [Nest
docs](https://docs.nestjs.com/graphql/other-features#execute-enhancers-at-the-field-resolver-level).