Fix drag and drop reordering in collection blocks (#1344)
The implementation of `moveBlock` was flawed. Moving blocks by swapping
only works if the move happens between directly adjacent blocks.
However, there are scenarios where the move should be performed between
non-adjacent blocks, for instance when a block is dragged outside the
block list (see screencasts). To fix this, we change the implementation
to move by removing the block at the `from` index and inserting it at
the `to` index.