CRUD Generator: Add support for integer IDs (#1320)
The `id` of an entity may be an integer. The CRUD generator previously assumed IDs to always be strings. This change adds support for integer IDs in all operations. When using an integer ID, `number` is used as type for the ID and a transform to convert the passed GraphQL argument to a number is added.