Prevent overlapping of clear-button with select-arrow
Since the end-adornment for the clear-button needs absolute
positioning, it is no longer possible to use the `clearable` prop in
combination with an  `endAdornment`. However, this is not a common
use-case, and an `endAdornment` on a select is generally not supported
by MUI, see: https://github.com/mui/material-ui/issues/17799.

An alternative would have been to position the select-arrow relative.
Unfortunately, this would cause the clickable element that opens the
select to shrink and prevent opening the select by clicking the
select-arrow. Since this would confuse users, the `endAdornment` is
now also positioned absolutely.