Fix copying non-image files (#1394)
Copying pages that used a non-image file failed. The reason was that the
`findCopiesOfFileInScope` query is used to check for copies of a file.
This query has a nullable imageCropArea argument. Still, the validation
failed when imageCropArea wasn't provided. The args of 
findCopiesOfFileInScope were extracted to a @ArgsType(). Now the 
validation works as expected.