COM-129: Fix Page Properties Dialog  (#1414)
## Previously:

The page properties dialog contains a slug field. This field has an
async validation to check in the API if the resulting path is still
available.

This led to a problem when saving the form in an `EditDialog`. In the
save action, the form was saved using `formRenderProps.form.submit()`.
However, this function doesn't wait for the validation and the internal
submit to be finished.

This led to a "Discard unsaved changes" dialog showing after
clicking save:

https://github.com/vivid-planet/comet/assets/13380047/4ff77bec-746c-46d3-b6ae-9e226ce8a324

## Now:

Now the submission is handled correctly. The save action waits for the
validation and submission to finish successfully before closing the
dialog.


https://github.com/vivid-planet/comet/assets/13380047/2fa7d2e6-830c-44a3-b2d4-0a540c04bf33