Fix fixtures (#1515)
## Problem:

Our fixtures are currently broken. You always get this error when
executing them:

<img width="573" alt="Bildschirmfoto 2023-12-14 um 17 02 52"
src="https://github.com/vivid-planet/comet/assets/13380047/ce7eed3c-5834-4618-8ced-ec93875c71ca">


The problem is that a required `userGroup` field is added to
`PageTreeNode` in Demo:


https://github.com/vivid-planet/comet/blob/99e0c57cc8ef542befd74d5b8ad9cf71058570e1/demo/api/src/page-tree/entities/page-tree-node.entity.ts#L26-L28


https://github.com/vivid-planet/comet/blob/1c098e4d0f23946c0fd543856c8f0aa897b83b82/demo/api/src/page-tree/dto/page-tree-node.input.ts#L7-L12

We don't provide a userGroup in the fixtures and it didn't strike
because the typing of `PageTreeService.createNode()` is wrong. The input
is typed with `PageTreeNodeBaseCreateInput`


https://github.com/vivid-planet/comet/blob/b73eb4c34c70c694fc8b267ce330af5959c50e7b/packages/api/cms-api/src/page-tree/page-tree.service.ts#L33

but the actual input type can be defined in the application and is
passed to the resolver via a factory:


https://github.com/vivid-planet/comet/blob/46bced34cbb9ec02de6dd462bd9111a6be7898b7/packages/api/cms-api/src/page-tree/createPageTreeResolver.ts#L35

### -> So actually the typing in the `PageTreeService` is wrong and we
need to fix that

I now circumvented this by using ts-ignore. 
But I think we need a solution to get the typing right. I'm however not
sure what the best way to do this is.