Fix: Enable copying documents from one `PageTree` category to another (#1519)
## Previously:

It wasn't possible to copy a document from one category to another (e.g.
from "Main navigation" to "Top menu"). Instead, the document was always
pasted to its original category because we always used the category of
the copied node in `sendPages()`.
 

https://github.com/vivid-planet/comet/assets/13380047/8112d742-f73c-47cb-a664-516ee4e3286b

## Now:

The active category is passed via the `PageTreeContext`. So the node is
always pasted to the active category.


https://github.com/vivid-planet/comet/assets/13380047/b947feb8-c9f5-4b94-816b-3f89641f31ce