Backport fix Page Properties Dialog (#1414) to v4 (#1353)
de facto duplicate of https://github.com/vivid-planet/comet/pull/1414

## Previously:

The page properties dialog contains a slug field. This field has an
async validation to check in the API if the resulting path is still
available.

This led to a problem when saving the form in an `EditDialog`. In the
save action, the form was saved using `formRenderProps.form.submit()`.
However, this function doesn't wait for the validation to be finished.

On main, this led to the dialog closing and the changes being omitted if
the save was triggered while the validation request was still pending.


https://github.com/vivid-planet/comet/assets/13380047/58891028-fc79-460d-b9bf-e48c52587ef9

## Now:

Now the submission is handled correctly. The save action waits for the
validation and submission to finish successfully before closing the
dialog.


https://github.com/vivid-planet/comet/assets/13380047/2fa7d2e6-830c-44a3-b2d4-0a540c04bf33