Add `SaveBoundary` to allow saving multiple things (e.g., forms) with a centralized save button (#1448)
This introduces a SaveBoundary (context) that can contain multiple areas
that get saved together using a SaveBoundarySaveButton.

Any component can register itself in the SaveBoundary by providing a
hasChanges boolean and a doSave method. This also adds a default
implementation for FinalForm:
- if FinalForm is rendered inside a SaveBoundary, it registers
- else the FinalForm renders (like previously) a router Prompt

Prompts are not needed for the inner components, instead SaveBoundary
renders a single prompt containing all hasChanges from registered
components.

See products admin for an implementation example.

Also changed in Demo Products: ProductsPage now contains Toolbar and
Save Button for Detail Page. That way the form is now completely
independent of where it is used and could also be put into an
EditDialog.

#### Alternative:

A possible alternative would be to re-use the router Prompt component
that also has a save action. But I decided for an additional
implementation because:
- PromptHandler stores dirty in a ref instead of state, so
enabling/disabling save button is not possible (could be changed to
state though)
- Prompt has no clear edges, as a new SubmissionBoundary has

---

see #1449 for additional usage of SubmissionBoundary

---------

Co-authored-by: Johannes Obermair <48853629+johnnyomair@users.noreply.github.com>