Extend demo with manufacturer (#1729)
for a usecase with double-nested fields. Added second version using
embeddable as this should also be considered/tested in some cases. I'm
not sure if the naming should be "real" (like renaming
addressAsEmbeddable e.g. "officeAddress") as the idea behind the second
version is to test the same structure but with a different technique.

```
query QueryManufacturer {
  manufacturer(id:"8251c3ad-07f4-4493-9d5b-42fa7ea73b98") {
    id
    address {
      street
      streetNumber // optional
      zip
      country
      alternativeAddress { // optional
        street
        streetNumber // optional
        zip
        country
      }
    }
    addressAsEmbeddable {
      street
      streetNumber // optional
      zip
      country
      alternativeAddress {
        street
        streetNumber // optional
        zip
        country
      }
    }
  }
}
```