Execute JSON.parse() twice to parse escaped JSON string in `KubernetesService#getContentScope` (#1778)
Fix parsing of `contentScopeAnnotation` in
`KubernetesService#getContentScope`

---

You can see that only `json2` (after the second JSON.parse()) is an
object

<img width="572" alt="Bildschirmfoto 2024-02-28 um 15 14 47"
src="https://github.com/vivid-planet/comet/assets/13380047/668b073e-226e-497b-a4f7-894a1372c016">

---

This bug is new in v6, but I don't know yet why it occurs now. This is a
workaround so updating to v6 is possible for content websites. I'll
still try to figure out the reason so maybe we can remove this again in
the future.