Use single site instance for Previews (#1596)
We now expect the SitePreview to be protected with a login (Url is set
via SITE_PREVIEW_URL). The site-preview is invoked via opening an
api-route. From there we make a call to the API which validates the
currentUser of the protected Site-Preview against the requested
content-scope. If the validation is successful the current content-scope
can be stored for further usage (not done here as the contentScope is
static).

The block-preview can use the same (protected) instance, however, there
is no explicit check. The content scope is stored IFrameBridge for
further usage.

Another PR is planned which consolidates the storage and usage of the
current content scope.