Site loader: add caching mechanismn to requests made in block loader (#1682)
This adds a simple cache for block loader in block preview. The cache is
client side and in block preview only. No clearing, garbage collection
or so are needed as it lives only in the preview page in the site
iframe.

Pro:
- simple to use, no new api

Con:
- only works for graphql, adds hard dependency on graphql-request to
`@comet/cms-site`

TODO:

- [ ] move re-usable parts into library (which would probably cause a
hard dependency on graphql-request)

alternative to #1681 which adds a cache a level above.