Demo: Implement product variants editing in product form (#1727)
Add product variants editing using a grid and nested form in the
"Variants" tab of the product form. Demonstrates how to achieve a
central save button in combination with a tab that contains a grid.

Also adds a new `StackToolbar` component, which is a thin wrapper around
`Toolbar` to only show the last toolbar inside a stack. Otherwise two
toolbars would be shown in the product variants form.

Co-authored-by: Johannes Obermair <48853629+johnnyomair@users.noreply.github.com>