Automatically generate `className` for component slots with `createComponentSlot()` (#1809)
E.g. the `Root` slot of `MainContent` will now alway the className
`CometAdminMainContent-root`.
If the `fullHeight` prop is set to true, the className
`CometAdminMainContent-fullHeight` also will be added.

This was previously done automatically, when using `withStyles()` from
the legacy `@mui/styles` package, which was removed during the theming
refactor: #1376

This now matches the behavior of Comet-Admin components before the
theming refactor and MUI components.