Create rewrites for redirect loop (#1787)
Creates rewrites for redirect loop, as case sensitivity in Next.js is
inconsistent (see also
https://github.com/vercel/next.js/issues/21498#issuecomment-848315717).

--- 

Background:

Sometimes you want a redirect from an uppercase to a lowercase URL (e.g.
`/Example` to `/example`) for SEO purposes. This can be created in our
admin without problems. The casing is saved correctly in the DB and
transferred to Next.

The problem is that this creates a redirect loop in the site. The reason
is that the case sensitivity of Next.js is inconsistent. The delivery of
pages is case sensitive. The redirects are case insensitive:

Without redirect:

http://localhost:3000/example -> page is delivered
http://localhost:3000/Example -> 404

With redirect:

http://localhost:3000/example -> Redirect to /example -> Loop
http://localhost:3000/Example -> Redirect to /example -> Loop

Also described here: https://github.com/vercel/next.js/issues/21498

---------

Co-authored-by: Thomas Dax <thomas.dax@vivid-planet.com>