Change `DatePicker` and `DateRangePicker` values from `Date` to `string` (#1916)
The reason for this change is that when selecting a date like
`2024-04-10` in a timezone ahead of UTC, it would be stored in a `Date`
object as e.g. `2024-04-09T22:00:00.000Z`. When only the date is saved
to the database, without the time, it would be saved as `2024-04-09`,
which differs from the selected date.