Admin Generator (Future): Don't generate toolbar in form, should be added in parent page component (#1937)
And adjust demo Products Future.

This is a major braking change, but we never released a stable version
containing the future generator.

## Alternative

we could add a configuration if the toolbar should be generated. But I
think we usually should not do that as we always want to use the
SaveBoundary.