Hide the "Dependents" tab in the DAM if the `DependenciesConfigProvider` is not configured (#1956)
Previously, the tab was always shown, even if the feature wasn't
configured. Though it didn't cause an error, the tab showed no valuable
information. This basically forced projects to configure the feature or
risk bad UX.

Now, we hide the tab completely if no configuration is passed via the
`DependenciesConfigProvider`.