Fix scope check for DAM bulk actions (#1976)
Currently, the mutations `moveDamFiles`, `copyFilesToScope`,
`archiveDamFiles` and `restoreDamFiles` all cause an error because they
don't have `@AffectedEntity` or `skipScopeCheck`.

This PR adds `@AffectedEntity` to all of them and removes previously
added custom scope validation logic. It further adds a `nullable` param
to `@AffectedEntity` for idArgs that can be null or undefined.