Require `aspectRatio` prop for `PixelImageBlock` and `Image` (#2012)
We repeatedly have the problem that an image is not displayed in the
site as the customer would expect. It is then claimed that our smart
focus point feature doesn't work properly. In most cases, however, the
error is due to an incorrectly used image in the site. For instance, an
image that should fill a mobile screen (roughly 1x2) may be used with
`layout="fill"` but specifying no aspect ratio, which then results in
using the default aspect ratio 16x9.

I suspect that these problems could be solved by enforcing the aspect
ratio for the `PixelImageBlock` and the `Image` component in the site.
This would force our frontend devs to pause and think about which aspect
ratio to use every time they use an image block or image.