Fix API Generator Linting (#2033)
## Previously

While linting the generated files, an error like this occured:

```
Parsing error: ESLint was configured to run on `` using `parserOptions.project`: \n' +
      'However, that TSConfig does not include this file. Either:\n' +
      "- Change ESLint's list of included files to not include this file\n" +
      '- Change that TSConfig to include this file\n' +
      '- Create a new TSConfig that includes this file and include it in your parserOptions.project\n'
```

This resulted in unformatted files, causing linting errors later on.

The TSConfig is needed by the `@typescript-eslint/naming-convention`
rule (see https://github.com/vivid-planet/comet/pull/1656).

## Reason

Apparently, the reason for this behavior was that the file wasn't
created before executing the lint. It seems like ESLint tried to get
type data for the file via TSConfig but couldn't because the file didn't
exist.

## Fix

Now the file is created, then the lint is executed, then the file is
overwritten with the formatted code.