Add custom action to handmade products grid (#2044)
We need to manually set the width of the actions column, depending on
the number of `IconButton` components shown, 32px per `IconButton`.
DataGrid columns do not scale automatically, depending on their content.

We cannot use `getActions` from the [special property
types](https://mui.com/x/react-data-grid/column-definition/#special-properties)
for this:

- Every item returned by `getActions` must be the `GridActionsCellItem`
component directly. This would not allow us to wrap the item with a
`Tooltip` or add a Fragment that contains the item so we can add a
`Dialog` that opens when clicking the item.

- Additionally, `CrudContextMenu` is not currently implemented with
`GridActionsCellItem` so it cannot be used inside `getActions` in it's
current state. Simply replacing the items with `GridActionsCellItem`
would not work, as `getActions` requires an array of items and cannot
handle multiple items inside one element or fragment.