Adjust `searchToMikroOrmQuery` function to reduce the amount of irrelevant results (#2075)
This is done by a combination of an and- & or-query.
It makes a search more restrictive. For example a search of `Red Shirt`
won't give all products containing `red` or `shirt` but rather returns
all products that have the words `red` and `shirt` in some column. The
words don't have to be in the same column.

---------

Co-authored-by: Johannes Obermair <48853629+johnnyomair@users.noreply.github.com>