Fix Drag and Drop on mobile devices (#2204)
This is done by adding a [multi-backend DnD
provider](https://www.npmjs.com/package/react-dnd-multi-backend?activeTab=readme)
and a [html5 to touch pipeline](rdndmb-html5-to-touch) package. Both are
used in the `App.tsx`. The result was tested in normal chrome, the
chrome mobile mode, on an android smartphone on chrome and in Safari on
both an iPhone and an iPad.