Lint generated schema files in API packages (#2222)
The GraphQL schema and block meta generation in API packages repeatedly
didn't work due to missing and/or broken imports. To prevent this, we
generate the files in the lint workflow, where an error during
generation should be more visible. This furthermore allows as to verify
that the files are up-to-date (similar to the API Generator lint).