Admin Stack: Update parent breadcrumb URL in state to not forget filters and other states when going back (#2159)
This adds to Stack a mechanism that watches the current location and
updates the last breadcrumb item with the new url. This fixes back
button (or any breadcrumb link) to go to the expected, previously open
url, to preserve states that are stored in the url - such as filter in
grid:

[Bildschirmaufzeichnung vom 2024-06-10,
12:59:01.webm](https://github.com/vivid-planet/comet/assets/50764/92feecb7-60a9-4170-b771-d61a3735accf)

This PR removes "invisible" Breadcrumbs (which is theoretically an
incompatible change) which where used (only) in RouterTabs - to support
a similar feature like this PR - but every component (as RouterTabs) had
to add support for it. The new solution doesn't need that as it watches
the location. Drawback: this doesn't work when refreshing, then the
stored state is lost.

---------

Co-authored-by: Johannes Obermair <48853629+johnnyomair@users.noreply.github.com>