Remove ESLint rule for enum naming (#2266)
I left the `@typescript-eslint/prefer-enum-initializers` rule in place
because I think this one still makes sense

## Motivation: 

We tried to enforce PascalCase for enums, then camelCase for enums. In
both cases, there were some edge cases that are cumbersome to fix. So we
decided to use camelCase if possible from now on, but not enforce it via
ESLint.