Remove next/link legacy behavior (#2271)
This PR only demonstrates the necessary code changes. I will add a
migration guide and changesets once we decide we want to change that
now.

Changes necessary in projects:

Either: 
- All custom link blocks need to render a HTML tag (a or span) that
accepts a className
- All link stylings need to be change to style the link block, not the
a-tag in the link block
- className needs to be prop drilled in LinkBlock

This will require lots of work in the projects I suppose. The [Next
codemod
](https://nextjs.org/docs/app/building-your-application/upgrading/codemods#remove-a-tags-from-link-components)
will not be of much help here.

Or: Add `legacyBehavior` prop to link block usages.

---

<!-- Everything below this line will be removed from the commit message
when the PR is merged -->

## PR Checklist

- [x] Verify if the change requires a changeset. See
[CONTRIBUTING.md](https://github.com/vivid-planet/comet/blob/HEAD/CONTRIBUTING.md)
-   [x] Link to the respective task if one exists: COM-803

---------

Co-authored-by: Niko Sams <niko.sams@gmail.com>