API Generator: Add `search` argument to `OneToManyFilter` and `ManyToManyFilter` (#2238)
This PR adds search filter to OneToMany and ManyToMany relations, that
can be used where a select is impossible to use because of too many
related entities. Doing a poor-mans fulltext search doesn't have this
problem (also has poor performance on the server side though)

The old method (how search was implemented for list queries) didn't
apply here well, so I decided do go a different approach: Instead of
generating code (the service) that lists all fields to query, we find
those fields at runtime now. We are using (1) MikroORM metadata and (2)
our CrudField resolver, the search boolean. Just as before.

And with this new approach it is possible to do that also for related
entities that makes implementing the search possible.

---------

Co-authored-by: Johannes Obermair <johannes.obermair@vivid-planet.com>